requirements.txt
debugpy==1.8.14
pillow==11.2.1
PyAutoGUI==0.9.54
pyinstaller==6.13.0
PyYAML==6.0.1
PyQt5
requests==2.31.0
selenium==4.15.2
jinja2==3.1.2
MarkupSafe==2.1.3
opencv-python-headless==4.11.0.86
websocket-client==1.8.0